Initial input files, run simulation script, and test
This commit includes the input files for the different simulation
stages and phases, as well as the topology and configuration files for
the protein and ligand complexes. All input files were thereby broken
down as much as possible, to detail the similarities between the
simulations and simplify maintenance.

This commit also includes a run_simulation script which helps to
combine the input files for the different simulation stages and phases
and runs simulations. This script is tested by a dedicated test
script, which is also called on every push and pull request to main
via GitHub actions. This commit also contains the set up for this
action, including a conda environment and a workflow.

Finally, this updates the README file with an initial documentation of
the repository and how to use it.