Avoid using deprecated GROMACS option in test (#3)
The test was using the command line flag `-nsteps 2` to reduce the
number of steps `gmx mdrun` simulates. This allows all simulations to
run in the matter of seconds. This command line flag is, however,
deprecated.

This commit changes the test to actually modify the mdp files before
running the test. This is less elegant, but has two advantages:

1. It doesn't use a deprecated interface
2. It is more versatile, we will be able to change other settings in
   the same manner.