NES analysis scripts (#10)
* NES analysis scripts

This introduces the analysis functions to calculate the free energy from
a swarm of NES simulations. They have been carefully checked to reproduce
the values in the pre-publication.

It includes the analysis functions in a folder structure that can be used
to make this an installable package in the future.

The analysis functions include unit and regression tests, which are run
on every push and PR to main by Github actions. This also introduces some
basic linting for Python files to be run by Github action.